+++ app/main.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI, Request
 from fastapi.responses import HTMLResponse
 from fastapi.staticfiles import StaticFiles
 
-from .db import (
+from .store import (
     PlacedItemTable,
     PlacementTable,
     ProductTable,
