+++ app/store/placed_item.py
