+++ app/store/placement.py
